Find the root identifier of a CRANT neuron

Find the root identifier of a CRANT neuron

Usage

crant_rootid(x, integer64 = FALSE, ...)

Arguments

x

One or more FlyWire segment ids

integer64

Whether to return ids as integer64 type (more compact but a little fragile) rather than character (default FALSE).

...

Additional arguments passed to pbapply::pbsapply and eventually to Python cv$CloudVolume object.

Value

A vector of root ids (by default character)

See also

Examples

# \donttest{
crant_rootid("576460752684030043")
#> Error : Unable to find chunked graph credentials!
#> Please set by doing:
#>   flywire_set_token()/fanc_set_token() etc 
#> For further details see:
#> https://github.com/seung-lab/cloud-volume#chunkedgraph-secretjson
#> Error in crant_scene(): It looks like you do not have a stored token. Please use `flywire_set_token()` to make one.Note you may have to do this in addition to `crantr_set_token()`
# }